fix(widgets): rename data.tsx
