+++ src/widgets/faq/model/meta.tsx
